Silly Kent, the idea was to have the status windows slide out from the right side of the screen, until the right most edge of said status window is 32 pixels away from the right edge of the screen. Your code doesn't do that. Your code slides it out from the right edge of the screen, until the left edge of the status window hits the exact middle of the screen.
Bad, bad, bad, BAD Kent.

Anyway, Here's code that actually works! Took some work, and it might not
look the best, so feel free to update it's syntax (just make sure that
the outcome is the same, otherwise you die).

Oh, yeah, don't use NSLogs in things you commit. That's what ITDebugLog()
is for, if anything.